tinydtls tests: Fixes for Clang
Clang is a bit stricter when it comes to static arrays with the size of
a static const, replacing with defined expressions.

The removed allowed compiler warnings were only relevant to the older
released versions of tinydtls ("master") branch, which was never used in
a committed version of this port -- but the new ones are required for
Clang and do not disturb GCC operation.